code reorganisation to move / rename remote index helpers out of RPC helpers module
no code changes
