+++ happs/observation/zomes/process/code/src/process_requests.rs
@@ -19,9 +19,9 @@ use hdk_graph_helpers::{
         query_direct_index_with_foreign_key,
         query_direct_remote_index_with_foreign_key,
     },
-    rpc::{
+    remote_indexes::{
         RemoteEntryLinkResponse,
-        handle_remote_index_sync_request,
+        handle_sync_direct_remote_index_destination,
     },
 };
 
@@ -94,7 +94,7 @@ pub fn receive_query_processes(params: QueryParams) -> ZomeApiResult<Vec<Respons
 }
 
 pub fn receive_link_committed_inputs(base_entry: CommitmentAddress, target_entries: Vec<ProcessAddress>, removed_entries: Vec<ProcessAddress>) -> ZomeApiResult<RemoteEntryLinkResponse> {
-    handle_remote_index_sync_request(
+    handle_sync_direct_remote_index_destination(
         COMMITMENT_BASE_ENTRY_TYPE,
         COMMITMENT_INPUT_OF_LINK_TYPE, COMMITMENT_INPUT_OF_LINK_TAG,
         PROCESS_COMMITMENT_INPUTS_LINK_TYPE, PROCESS_COMMITMENT_INPUTS_LINK_TAG,
@@ -103,7 +103,7 @@ pub fn receive_link_committed_inputs(base_entry: CommitmentAddress, target_entri
 }
 
 pub fn receive_link_committed_outputs(base_entry: CommitmentAddress, target_entries: Vec<ProcessAddress>, removed_entries: Vec<ProcessAddress>) -> ZomeApiResult<RemoteEntryLinkResponse> {
-    handle_remote_index_sync_request(
+    handle_sync_direct_remote_index_destination(
         COMMITMENT_BASE_ENTRY_TYPE,
         COMMITMENT_OUTPUT_OF_LINK_TYPE, COMMITMENT_OUTPUT_OF_LINK_TAG,
         PROCESS_COMMITMENT_OUTPUTS_LINK_TYPE, PROCESS_COMMITMENT_OUTPUTS_LINK_TAG,
@@ -112,7 +112,7 @@ pub fn receive_link_committed_outputs(base_entry: CommitmentAddress, target_entr
 }
 
 pub fn receive_link_intended_inputs(base_entry: IntentAddress, target_entries: Vec<ProcessAddress>, removed_entries: Vec<ProcessAddress>) -> ZomeApiResult<RemoteEntryLinkResponse> {
-    handle_remote_index_sync_request(
+    handle_sync_direct_remote_index_destination(
         INTENT_BASE_ENTRY_TYPE,
         INTENT_INPUT_OF_LINK_TYPE, INTENT_INPUT_OF_LINK_TAG,
         PROCESS_INTENT_INPUTS_LINK_TYPE, PROCESS_INTENT_INPUTS_LINK_TAG,
@@ -121,7 +121,7 @@ pub fn receive_link_intended_inputs(base_entry: IntentAddress, target_entries: V
 }
 
 pub fn receive_link_intended_outputs(base_entry: IntentAddress, target_entries: Vec<ProcessAddress>, removed_entries: Vec<ProcessAddress>) -> ZomeApiResult<RemoteEntryLinkResponse> {
-    handle_remote_index_sync_request(
+    handle_sync_direct_remote_index_destination(
         INTENT_BASE_ENTRY_TYPE,
         INTENT_OUTPUT_OF_LINK_TYPE, INTENT_OUTPUT_OF_LINK_TAG,
         PROCESS_INTENT_OUTPUTS_LINK_TYPE, PROCESS_INTENT_OUTPUTS_LINK_TAG,
